Make backend visibility check work on all browsers
As noted in the TODO this wasn't working on Safari because it
used the not 100% supported (Safari, IE, Samsung) Element visibility
DOM API. https://caniuse.com/mdn-api_element_checkvisibility

We can always switch to it if/when Safari is back in the fold but it's
not a very tricky check anyway.

I used https://stackoverflow.com/a/21696585/385622 as a reference,
caveat being it can mistakenly count `position: fixed` elements
as being invisible (because they don't have an `offsetParent`).

That doesn't curently apply to the `backends_list` element here, but
it's good to keep in mind.